feat: add auto_controller method to Controller class
The auto_controller method is added to the Controller class in order to control the Marty robot's movements automatically. The method takes an argument NBMarty to specify which Marty robot to control. If NBMarty is 1, the method updates the tab1 list with the Marty robot's movements. If NBMarty is 2, the method updates the tab2 list. Otherwise, an error message is printed.

Co-authored-by: aa14728 <arthur_albert@etu.u-bourgogne.fr>