feat: Update auto_controller method to include color sensor readings
The auto_controller method in the Controller class is updated to include color sensor readings, providing more information about the environment. The method now takes an argument NBMarty to specify which Marty robot to control. If NBMarty is 1, the method updates the tab1 list with the Marty robot's movements. If NBMarty is 2, the method updates the tab2 list. Otherwise, an error message is printed.