feat: Improve auto_controller method in Controller class
The auto_controller method in the Controller class is updated to improve the control of the Marty robot's movements. Color sensor readings are now included to provide more information about the environment. The method takes an argument NBMarty to specify which Marty robot to control. If NBMarty is 1, the method updates the tab1 list with the Marty robot's movements. If NBMarty is 2, the method updates the tab2 list. Otherwise, an error message is printed.